src/requirements.txt
streamlit==1.45.1
streamlit-authenticator==0.4.2
python-dotenv==1.1.0
openai==1.56.2
azure-identity==1.19.0
tiktoken==0.9.0
st-copy==1.0.2
streamlit-mermaid==0.3.0